+++ cdmw/ui/shell/signal_wiring.py
@@ -40,7 +40,6 @@ class ShellSignalWiringMixin:
         self.archive_scan_button.clicked.connect(self.scan_archives)
         self.archive_refresh_scan_button.clicked.connect(lambda: self.scan_archives(force_refresh=True))
         self.archive_asset_catalog_button.clicked.connect(self._show_archive_asset_catalog_dialog)
-        self.archive_material_finder_button.clicked.connect(self._show_archive_material_finder_dialog)
         self.archive_clear_asset_scope_button.clicked.connect(self._clear_archive_asset_catalog_scope)
         self.archive_extract_selected_button.clicked.connect(self.extract_selected_archive_entries)
         self.archive_extract_filtered_button.clicked.connect(self.extract_filtered_archive_entries)